PR 119757: Merge brendyna/feature/getting-started to master
The purpose of this area is to create an easy-to-copy-and-reference sample for newcomers to this project.

Richard/Sarah/Kashif: if you could take a look and let me know if anything is missing that'd be helpful, I'd greatly appreciate it! You can see this by syncing my branch locally, hitting F5, then checking out the new, blue-highlighted row pointing to the Getting Started area. All the documentation is in that view. Let me know if you run into any issues.

 - Adding GettingStarted area and base files
 - Stripping down components to minimum required (and getting build working)
 - Moving IRepositoryParams to repositories and cleaning up unused interfaces
 - Removing example CSS from getting started (ideally, all CSS is deferred to
 - Adding Getting Started to index page and specially calling it out
 - Moving IRepositoryParams to repositories and removing unused interfaces (missed Filters repo)
 - Commenting base classes and ordering properties correctly
 - Adding anchor to Section subsection to support TOC linking to subsections
 - Adding remaining Getting Started content/code
 - Getting unit tests fleshed out and passing

Related work items: #9416837